+++ app/(dashboard)/lib/queries/projects/actions.ts
@@ -6,18 +6,31 @@ import { BUCKET_NAME } from "@/app/(dashboard)/constants/items.constants";
 import { IProject } from "@/app/(dashboard)/types/data";
 import { capitalize } from "@/lib/helpers";
 import { createClient } from "@/lib/supabase/server";
+import { translates } from "@/lib/translations";
+import { normalizeNodeId } from "platejs";
 import { ISortOrder } from "../../search-params";
+import readUserSession from "@/lib/read-session";
 
 const TABLE_NAME = "projects";
 
 const getSelectedProject = async (slug: IProject["slug"]) => {
   const supabase = await createClient();
 
-  const { data, error } = await supabase.from(TABLE_NAME).select("asset_id, title, slug, description, overview, tech_stack, urls").eq("slug", slug).single();
+  const { data, error } = await supabase
+    .from(TABLE_NAME)
+    .select(
+      "asset_id, title, slug, description, overview, tech_stack, roles, project_status, start_date, is_current, end_date, urls",
+    )
+    .eq("slug", slug)
+    .single();
 
   if (error) throw error;
 
-  const { data: asset } = await supabase.from(BUCKET_NAME).select("file_name").eq("id", data.asset_id).single();
+  const { data: asset } = await supabase
+    .from(BUCKET_NAME)
+    .select("file_name")
+    .eq("id", data.asset_id)
+    .single();
 
   const project = { ...data, file_name: asset?.file_name } as IProject;
 
@@ -29,7 +42,7 @@ const getFilteredProjects = async ({
   pageLimit = 10,
   category,
   search,
-  sort
+  sort,
 }: {
   page?: number;
   pageLimit?: number;
@@ -43,7 +56,8 @@ const getFilteredProjects = async ({
   const from = (page - 1) * pageLimit;
   const to = from + pageLimit - 1;
 
-  let query = supabase.from(TABLE_NAME).select(`
+  let query = supabase.from(TABLE_NAME).select(
+    `
     asset_id, 
     title, 
     slug,
@@ -55,14 +69,18 @@ const getFilteredProjects = async ({
     ),
     created_at,
     updated_at
-  `, { count: "exact" });
+  `,
+    { count: "exact" },
+  );
 
   if (search) {
     const searchText = search.trim();
     const upperSearchText = searchText.toUpperCase();
     const capitalizedSearchText = capitalize(searchText);
 
-    query = query.or(`title.ilike.%${searchText}%,tech_stack.cs.{"${searchText}"},tech_stack.cs.{"${upperSearchText}"},tech_stack.cs.{"${capitalizedSearchText}"}`);
+    query = query.or(
+      `title.ilike.%${searchText}%,tech_stack.cs.{"${searchText}"},tech_stack.cs.{"${upperSearchText}"},tech_stack.cs.{"${capitalizedSearchText}"}`,
+    );
   }
 
   if (category) {
@@ -98,10 +116,9 @@ const getFilteredProjects = async ({
       tech_stack: project.tech_stack,
       urls: project.urls,
       created_at: project.created_at,
-      updated_at: project.updated_at
+      updated_at: project.updated_at,
     };
   }) as unknown as IProject[];
-
 
   return {
     success: true,
@@ -109,42 +126,87 @@ const getFilteredProjects = async ({
     projects,
     page,
     pageLimit,
-    offset: from
+    offset: from,
   };
 };
 
 const bulkAddProjects = async (payload: TAddProjectsFormValues["projects"]) => {
+  const authUser = await readUserSession();
+  const role = authUser?.app_metadata?.role;
+
+  if (role !== "owner")
+    throw new Error("You're not authorized to perform this action");
+
   const supabase = await createClient();
 
-  const modifiedPayload = payload.map((project) => {
-    return {
+  const modifiedPayload = await Promise.all(
+    payload.map(async (project) => ({
       title: project.title,
       slug: project.slug,
       asset_id: project.thumbnail,
+      overview: normalizeNodeId(project.overview),
+      overview_id: await translates.gemini({
+        nodes: project.overview,
+        targetLang: "Bahasa Indonesia",
+      }),
       description: project.description,
-      overview: project.overview,
+      description_id: await translates.deepl({ texts: project.description }),
       tech_stack: project.tech_stack,
       urls: project.urls,
-    };
-  });
+    })),
+  );
 
-  const { error: insertError } = await supabase.from(TABLE_NAME).insert(modifiedPayload);
+  const { error: insertError } = await supabase
+    .from(TABLE_NAME)
+    .insert(modifiedPayload);
 
   if (insertError) throw insertError;
 };
 
-const updateSelectedProject = async (slug: IProject["slug"], payload: Partial<TUpdateProjectFormValues>) => {
+const updateSelectedProject = async (
+  slug: IProject["slug"],
+  payload: Partial<TUpdateProjectFormValues>,
+) => {
+  const authUser = await readUserSession();
+  const role = authUser?.app_metadata?.role;
+
+  if (role !== "owner")
+    throw new Error("You're not authorized to perform this action");
+
   const supabase = await createClient();
 
-  const updated_at = Date.now().toLocaleString();
-  const modifiedPayload = { ...payload, updated_at };
+  const updated_at = new Date().toISOString();
 
-  const { error } = await supabase.from(TABLE_NAME).update(modifiedPayload).eq("slug", slug);
+  const modifiedPayload = {
+    ...payload,
+    ...(payload.description && {
+      description_id: await translates.deepl({ texts: payload.description }),
+    }),
+    ...(payload.overview && { overview: normalizeNodeId(payload.overview) }),
+    ...(payload.overview && {
+      overview_id: await translates.gemini({
+        nodes: payload.overview,
+        targetLang: "Bahasa Indonesia",
+      }),
+    }),
+    updated_at,
+  };
+
+  const { error } = await supabase
+    .from(TABLE_NAME)
+    .update(modifiedPayload)
+    .eq("slug", slug);
 
   if (error) throw error;
 };
 
 const deleteSelectedProject = async (slug: IProject["slug"]) => {
+  const authUser = await readUserSession();
+  const role = authUser?.app_metadata?.role;
+
+  if (role !== "owner")
+    throw new Error("You're not authorized to perform this action");
+
   const supabase = await createClient();
 
   const { error } = await supabase.from(TABLE_NAME).delete().eq("slug", slug);
@@ -153,6 +215,12 @@ const deleteSelectedProject = async (slug: IProject["slug"]) => {
 };
 
 const deleteSelectedProjects = async (ids: IProject["id"][]) => {
+  const authUser = await readUserSession();
+  const role = authUser?.app_metadata?.role;
+
+  if (role !== "owner")
+    throw new Error("You're not authorized to perform this action");
+
   const supabase = await createClient();
 
   const { error } = await supabase.from(TABLE_NAME).delete().in("id", ids);
@@ -160,5 +228,11 @@ const deleteSelectedProjects = async (ids: IProject["id"][]) => {
   if (error) throw error;
 };
 
-export { bulkAddProjects, deleteSelectedProject, deleteSelectedProjects, getFilteredProjects, getSelectedProject, updateSelectedProject };
-
+export {
+  bulkAddProjects,
+  deleteSelectedProject,
+  deleteSelectedProjects,
+  getFilteredProjects,
+  getSelectedProject,
+  updateSelectedProject,
+};
